# contracts/openapi/simulation.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "油田微电网新能源协同控制仿真服务",
    "description": "单机离线模拟数据服务；不连接真实 D5000、SCADA 或现场设备。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/simulations/quick": {
      "post": {
        "summary": "Run Quick Simulation",
        "description": "同步运行一个快速单微网算例，用于首个真实端到端闭环。",
        "operationId": "run_quick_simulation_api_simulations_quick_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations": {
      "post": {
        "summary": "Create Simulation",
        "description": "创建独立进程仿真会话；同一时刻最多运行一个重型任务。",
        "operationId": "create_simulation_api_simulations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "507": {
            "description": "任务存储已达容量上限或无法完整统计，暂不受理新任务"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations/{simulation_id}": {
      "get": {
        "summary": "Get Simulation",
        "operationId": "get_simulation_api_simulations__simulation_id__get",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations/{simulation_id}/cancel": {
      "post": {
        "summary": "Cancel Simulation",
        "operationId": "cancel_simulation_api_simulations__simulation_id__cancel_post",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations/{simulation_id}/result": {
      "get": {
        "summary": "Get Simulation Result",
        "operationId": "get_simulation_result_api_simulations__simulation_id__result_get",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClusterTimeSeriesPoint": {
        "properties": {
          "time_hour": {
            "type": "number",
            "title": "Time Hour"
          },
          "aggregate_load_mw": {
            "type": "number",
            "title": "Aggregate Load Mw"
          },
          "aggregate_renewable_mw": {
            "type": "number",
            "title": "Aggregate Renewable Mw"
          },
          "aggregate_import_mw": {
            "type": "number",
            "title": "Aggregate Import Mw"
          },
          "aggregate_q_mvar": {
            "type": "number",
            "title": "Aggregate Q Mvar"
          },
          "import_limit_mw": {
            "type": "number",
            "title": "Import Limit Mw"
          },
          "regional_import_mw": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Regional Import Mw"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "time_hour",
          "aggregate_load_mw",
          "aggregate_renewable_mw",
          "aggregate_import_mw",
          "aggregate_q_mvar",
          "import_limit_mw",
          "regional_import_mw"
        ],
        "title": "ClusterTimeSeriesPoint"
      },
...
      "EventType": {
        "type": "string",
        "enum": [
          "pv_surge",
          "wind_surge",
          "load_drop",
          "load_surge",
          "communication_packet_loss",
          "communication_outage"
        ],
        "title": "EventType"
      },
      "ExecutiveSummary": {
        "properties": {
          "overall_passed": {
            "type": "boolean",
            "title": "Overall Passed"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "economic_improvement_percent": {
            "type": "number",
            "title": "Economic Improvement Percent"
          },
          "baseline_economic_cost_cny": {
            "type": "number",
            "title": "Baseline Economic Cost Cny"
          },
          "optimized_economic_cost_cny": {
            "type": "number",
            "title": "Optimized Economic Cost Cny"
          },
          "optimized_peak_import_mw": {
            "type": "number",
            "title": "Optimized Peak Import Mw"
          },
          "total_active_loss_mwh": {
            "type": "number",
            "title": "Total Active Loss Mwh"
          },
          "minimum_power_factor": {
            "type": "number",
            "title": "Minimum Power Factor"
          },
          "minimum_voltage_pu": {
            "type": "number",
            "title": "Minimum Voltage Pu"
          },
          "maximum_voltage_pu": {
            "type": "number",
            "title": "Maximum Voltage Pu"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "overall_passed",
          "headline",
          "economic_improvement_percent",
          "baseline_economic_cost_cny",
          "optimized_economic_cost_cny",
          "optimized_peak_import_mw",
          "total_active_loss_mwh",
          "minimum_power_factor",
          "minimum_voltage_pu",
          "maximum_voltage_pu"
        ],
        "title": "ExecutiveSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobState": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled",
          "interrupted"
        ],
        "title": "JobState"
      },
      "JobStatus": {
        "properties": {
          "simulation_id": {
            "type": "string",
            "title": "Simulation Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "stage_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Label"
          },
          "stage_sequence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Sequence"
          },
          "total_stages": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Total Stages",
            "default": 6
          },
          "result_available": {
            "type": "boolean",
            "title": "Result Available",
            "default": false
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "simulation_id",
          "name",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "JobStatus"
      },
...
      "ScenarioEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 48,
            "minLength": 1,
            "title": "Event Id"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "target": {
            "type": "string",
            "enum": [
              "SC",
              "YA_B",
              "YA_C"
            ],
            "title": "Target"
          },
          "time_axis": {
            "$ref": "#/components/schemas/EventTimeAxis"
          },
          "start": {
            "type": "number",
            "minimum": 0.0,
            "title": "Start"
          },
          "end": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "End"
          },
          "magnitude": {
            "type": "number",
            "maximum": 5.0,
            "exclusiveMinimum": 0.0,
            "title": "Magnitude",
            "default": 1.0
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Label",
            "default": "场景事件"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "event_type",
          "target",
          "time_axis",
          "start",
          "end"
        ],
        "title": "ScenarioEvent",
        "description": "Executable event with current iteration semantics."
      },
...
      "ScenarioType": {
        "type": "string",
        "enum": [
          "single_microgrid",
          "cluster_coordination",
          "communication_fault",
          "group_control"
        ],
        "title": "ScenarioType"
      },
      "SimulationMetadata": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.3.0"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "data_source": {
            "type": "string",
            "const": "parameterized_simulation",
            "title": "Data Source",
            "default": "parameterized_simulation"
          },
          "data_notice": {
            "type": "string",
            "title": "Data Notice",
            "default": "全部网络与运行数据均为参数化模拟数据，不代表实际油田数据。"
          },
          "scenario_type": {
            "$ref": "#/components/schemas/ScenarioType"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "steps": {
            "type": "integer",
            "title": "Steps"
          },
          "dt_hours": {
            "type": "number",
            "title": "Dt Hours"
          },
          "formulation": {
            "$ref": "#/components/schemas/SolverFormulation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "generated_at",
          "scenario_type",
          "region",
          "steps",
          "dt_hours",
          "formulation"
        ],
        "title": "SimulationMetadata"
      },
      "SimulationRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "default": "单微网快速仿真"
          },
          "scenario_type": {
            "$ref": "#/components/schemas/ScenarioType",
            "default": "single_microgrid"
          },
          "region": {
            "type": "string",
            "enum": [
              "SC",
              "YA_B",
              "YA_C"
            ],
            "title": "Region",
            "default": "SC"
          },
          "steps": {
            "type": "integer",
            "maximum": 288.0,
            "minimum": 1.0,
            "title": "Steps",
            "default": 8
          },
          "storage_enabled": {
            "type": "boolean",
            "title": "Storage Enabled",
            "default": true
          },
          "solver": {
            "$ref": "#/components/schemas/SolverOptions"
          },
          "admm_max_iterations": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 4.0,
            "title": "Admm Max Iterations",
            "default": 220
          },
          "communication_loss_probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Communication Loss Probability",
            "default": 0.05
          },
          "communication_max_delay_iterations": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Communication Max Delay Iterations",
            "default": 2
          },
          "communication_outage_region": {
            "type": "string",
            "enum": [
              "SC",
              "YA_B",
              "YA_C"
            ],
            "title": "Communication Outage Region",
            "default": "YA_B"
          },
          "communication_outage_start_iteration": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Communication Outage Start Iteration",
            "default": 12
          },
          "communication_outage_end_iteration": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Communication Outage End Iteration",
            "default": 22
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ScenarioEvent"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Events"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SimulationRequest",
        "description": "Executable request; validated separately from stored historical evidence."
      },
      "SimulationRequestRecord": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "default": "单微网快速仿真"
          },
          "scenario_type": {
            "$ref": "#/components/schemas/ScenarioType",
            "default": "single_microgrid"
          },
          "region": {
            "type": "string",
            "enum": [
              "SC",
              "YA_B",
              "YA_C"
            ],
            "title": "Region",
            "default": "SC"
          },
          "steps": {
            "type": "integer",
            "maximum": 288.0,
            "minimum": 1.0,
            "title": "Steps",
            "default": 8
          },
          "storage_enabled": {
            "type": "boolean",
            "title": "Storage Enabled",
            "default": true
          },
          "solver": {
            "$ref": "#/components/schemas/SolverOptions"
          },
          "admm_max_iterations": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 4.0,
            "title": "Admm Max Iterations",
            "default": 220
          },
          "communication_loss_probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Communication Loss Probability",
            "default": 0.05
          },
          "communication_max_delay_iterations": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Communication Max Delay Iterations",
            "default": 2
          },
          "communication_outage_region": {
            "type": "string",
            "enum": [
              "SC",
              "YA_B",
              "YA_C"
            ],
            "title": "Communication Outage Region",
            "default": "YA_B"
          },
          "communication_outage_start_iteration": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Communication Outage Start Iteration",
            "default": 12
          },
          "communication_outage_end_iteration": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Communication Outage End Iteration",
            "default": 22
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ScenarioEventRecord"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Events"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SimulationRequestRecord",
        "description": "Recorded inputs, including historical combinations no longer executable."
      },
      "SimulationResult": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/SimulationMetadata"
          },
          "request": {
            "$ref": "#/components/schemas/SimulationRequestRecord"
          },
          "executive_summary": {
            "$ref": "#/components/schemas/ExecutiveSummary"
          },
          "topology_nodes": {
            "items": {
              "$ref": "#/components/schemas/TopologyNode"
            },
            "type": "array",
            "title": "Topology Nodes"
          },
          "topology_edges": {
            "items": {
              "$ref": "#/components/schemas/TopologyEdge"
            },
            "type": "array",
            "title": "Topology Edges"
          },
          "validation_items": {
            "items": {
              "$ref": "#/components/schemas/ValidationItem"
            },
            "type": "array",
            "title": "Validation Items"
          },
          "comparison": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Comparison"
          },
          "solver": {
            "additionalProperties": true,
            "type": "object",
            "title": "Solver"
          },
          "timeseries": {
            "items": {
              "$ref": "#/components/schemas/TimeSeriesPoint"
            },
            "type": "array",
            "title": "Timeseries"
          },
          "cluster_timeseries": {
            "items": {
              "$ref": "#/components/schemas/ClusterTimeSeriesPoint"
            },
            "type": "array",
            "title": "Cluster Timeseries"
          },
          "admm_history": {
            "items": {
              "$ref": "#/components/schemas/ADMMHistoryPoint"
            },
            "type": "array",
            "title": "Admm History"
          },
          "communication": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CommunicationSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "group_control": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupControlSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "economic_accounting_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Economic Accounting Version"
          },
          "reference_economics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReferenceEconomics"
              },
              {
                "type": "null"
              }
            ]
          },
          "cluster_validation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClusterValidation"
              },
              {
                "type": "null"
              }
            ]
          },
          "coordination_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CoordinationSnapshot"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metadata",
          "request",
          "executive_summary",
          "topology_nodes",
          "topology_edges",
          "validation_items",
          "comparison",
          "solver",
          "timeseries"
        ],
        "title": "SimulationResult"
      },
      "SolverFormulation": {
        "type": "string",
        "enum": [
          "misocp",
          "legacy_milp"
        ],
        "title": "SolverFormulation"
      },
      "SolverOptions": {
        "properties": {
          "formulation": {
            "$ref": "#/components/schemas/SolverFormulation",
            "default": "misocp"
          },
          "time_limit_seconds": {
            "type": "number",
            "maximum": 1800.0,
            "exclusiveMinimum": 0.0,
            "title": "Time Limit Seconds",
            "default": 60.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SolverOptions"
      },
...
      "TopologyEdge": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "resistance_pu": {
            "type": "number",
            "title": "Resistance Pu"
          },
          "reactance_pu": {
            "type": "number",
            "title": "Reactance Pu"
          },
          "capacity_mva": {
            "type": "number",
            "title": "Capacity Mva"
          },
          "branch_kind": {
            "type": "string",
            "enum": [
              "line",
              "transformer"
            ],
            "title": "Branch Kind",
            "default": "line"
          },
          "tap_ratio": {
            "type": "number",
            "title": "Tap Ratio",
            "default": 1.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source",
          "target",
          "resistance_pu",
          "reactance_pu",
          "capacity_mva"
        ],
        "title": "TopologyEdge"
      },
      "TopologyNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "kind": {
            "type": "string",
            "enum": [
              "pcc",
              "main",
              "wind",
              "pv",
              "flex"
            ],
            "title": "Kind"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "kind"
        ],
        "title": "TopologyNode"
      },
...
    }
  }
}